requirements.txt
pyflink~=1.0
kafka-python~=2.0.2
python-dotenv~=1.0.1
env~=0.1.0
apache-flink~=1.20.0